Remove redundant Generator abstraction
